+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add SimpleList<T>
+
+### Changed
+- Optimize MaxFlow
+
 ## [1.0.8] - 2020-01-05
 ### Added
 - Add IShiftOperator interface